John

+++ GameHubApp/app/src/main/java/com/gamehub/myapplication/games.java
@@ -10,18 +10,18 @@ import java.util.*;
 import java.util.List;
 import java.util.UUID;
 public class games {
-    private Double gameID;
+    private Integer gameID;
     private String title;
     private String publisher;
-    private Integer rating;
+    private Double rating;
     // Should I have a slot for avg rating and also make a list of all of its ratings maybe mapped to user id so you cant have more than one vote
-    public games(Double gameID, String title, String publisher, Integer rating){
+    public games(Integer gameID, String title, String publisher, Double rating){
         this.gameID = gameID;
         this.title = title;
         this.publisher = publisher;
         this.rating = rating;
     }
-    public void setID(Double gameID){
+    public void setID(Integer gameID){
         this.gameID = gameID;
     }
     public void setTitle(String title){
@@ -30,10 +30,10 @@ public class games {
     public void setPublisher(String publisher){
         this.publisher = publisher;
     }
-    public void setRating(Integer rating){
+    public void setRating(Double rating){
         this.rating = rating;
     }
-    public Double getID(){
+    public Integer getID(){
         return this.gameID;
     }
     public String getTitle(){
@@ -42,7 +42,7 @@ public class games {
     public String getPublisher(){
         return this.publisher;
     }
-    public Integer getRating(){
+    public Double getRating(){
         return this.rating;
     }
 }
